refactor: 업로드 횟수 검증 Facade에서 호출로 변경
