+++ src/main/java/com/birca/bircabackend/command/cafe/presentation/BusinessLicenseController.java
@@ -3,8 +3,6 @@ package com.birca.bircabackend.command.cafe.presentation;
 import com.birca.bircabackend.command.auth.authorization.LoginMember;
 import com.birca.bircabackend.command.auth.authorization.RequiredLogin;
 import com.birca.bircabackend.command.cafe.application.BusinessLicenseFacade;
-import com.birca.bircabackend.command.cafe.application.BusinessLicenseProcessingService;
-import com.birca.bircabackend.command.cafe.application.BusinessLicenseService;
 import com.birca.bircabackend.command.cafe.dto.BusinessLicenseCreateRequest;
 import com.birca.bircabackend.command.cafe.dto.BusinessLicenseResponse;
 import lombok.RequiredArgsConstructor;
@@ -21,8 +19,6 @@ import org.springframework.web.multipart.MultipartFile;
 public class BusinessLicenseController {
 
     private final BusinessLicenseFacade businessLicenseFacade;
-    private final BusinessLicenseService businessLicenseService;
-    private final BusinessLicenseProcessingService businessLicenseProcessingService;
 
     @PostMapping("/v1/cafes/license-read")
     @RequiredLogin
@@ -37,8 +33,7 @@ public class BusinessLicenseController {
     @RequiredLogin
     public ResponseEntity<Void> saveBusinessLicense(LoginMember loginMember,
                                                     @ModelAttribute BusinessLicenseCreateRequest request) {
-        businessLicenseProcessingService.verifyBusinessLicenseStatus(request.businessLicenseNumber());
-        businessLicenseService.saveBusinessLicense(loginMember, request);
+        businessLicenseFacade.saveBusinessLicense(loginMember, request);
         return ResponseEntity.ok().build();
     }
 }
